requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.2
langchain==0.0.350
langchain-google-genai==0.0.5
google-generativeai==0.3.2
pandas==2.1.3
pypdf2==3.0.1
pdfplumber==0.10.2
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
uuid==1.30
python-dateutil==2.8.2 